Update Buyer.java
Changed the key from "UserType" to "User" to not make it be confused by the UserType interface.Also, I think to make the Strategy pattern more relevant, we shouldn't have to have the userEnum as a parameter. Buyer knows it's a buyer so it should just input User.BUYER directly into the intent. but I won't change it here until we speak.