ProductArrayAdapter - Added the ShoppingCart ListView stuff. Added a new constructor that doesn't have the User enum. Has spinners and such. I don't know why but it seems like the ListView adapter loops a few times before settling on a view.
ShoppingCart - Added listeners and such for the buttons. PayNow is still not implemented.

BuyerClerk - Implemented methods involving _shoppingCart and the Cart  class.

Cart - Implemented a way to keep track of the quantity of each product in the cart. Slightly crude but gets the job done. Open to suggestions on this one. Haven't implemented getPrice yet.

Product - Added a .equals() method to compare between Products.

custom_list_shopping_cart - Created a custom list view based off of custom_list_buyer/seller

shopping_cart - Implemented the UI with a button and ListView.